refactor(menu/spectate): general cleanup + cycling feature
This cleanups up our spectate system a little bit to hopefully help with readability and logic flow. 

But, much of it still relies heavily on side effects that make me weep :(

This also adds the ability cycle between players as a feature, so users wouldn't have to click on each individual player in order to spectate them.