feat(core): improve handling of checkJoin
Previously the bot was hitting the discord api every time a non-member
tries to join a server with the discord member whitelist enabled.
Now it will only try to re-fresh the cache every 60 seconds.

In this update we also gave more time for the bot to resolve the member
by changing the timings in sv_main.lua.